feat: split out api references from the view
